Add method that extracts yaml metadata and file body into separate strings
Needed this because Kramdown was attempting to convert the YAML metadata into HTML which I did not want. Instead, we can use this method to extract the file content that we want to pass to Kramdown. And then if we need to do something more with the YAML metadata later we can easily access it.